==> Rmd/07_kPCA_and_kPCR.Rmd <==
---
title: "07_kPCA_and_k_PCR"
author: "Callum Weinberg"
date: "May 21, 2022"
output:
  html_document:
    df_print: paged
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

# This is necessary given the nested directories.
# Enter your working directory here
knitr::opts_knit$set(root.dir = "/home/clw/Documents/UCSB/Quarter7/PSTAT 262FE/Final Project/PSTAT_262FE_Final")
```

## Libraries

```{r libraries}
library(kernlab)
library(ggplot2)
library(dplyr)
library(tidyr)
library(stringr)
library(cowplot)
```


## Load and Prepare Data

```{r load_data}
load(file="Intermediate/model_data.Rdata")

# Create Matrices
X_Matrix = model_data[,2:ncol(model_data)]
Y_Matrix = as.matrix(model_data[,1])

# Center and Scale Data
X_scale_center = scale(X_Matrix,center = TRUE, scale = TRUE)
```

## Define Gram Matcies

```{r gram_matrices}
# Euclidean Distance Matrix Formula
#RBF: Define Euclidean Distance Matrix 
euclidean_distance_function = function(M) {

  # Define Result Matrix for Outer Product
  result = matrix(NA,nrow = nrow(M),ncol = nrow(M))
  
  for(i in 1:nrow(M)) {
    for(j in 1:nrow(M)) {
      # Note: Taking sqrt to make proper
      # euclidean distance. Some kernels (rbf and 
      # rq) then square the terms.
      # Could decrease computing time slightly
      # By not calculating true euclidean
      # distance here
      result[i,j] = sqrt(sum((M[i,] - M[j,])^2))
    }
  }
  # Return Result
  eu_matrix = result
  return(eu_matrix)
}


#Radial Basis Function Kernel
rbf_kernel = function(M, l) {

  # The euclidean distance matrix is squared
  # then multiplied by -sigma and exponentiated
  gram = exp(-1*(euclidean_distance_function(M)^2)/l)
  return(gram)
}

# Rational Quadratic Kernel
rq_kernel = function(M,l,alpha) {
  gram = (1 + ((euclidean_distance_function(M)^2)/(l*alpha)))^(-1*alpha)
  return(gram)
}

# Periodic Kernel
per_kernel = function(M,l,p) {
  gram = exp((-1)*sin(abs(euclidean_distance_function(M))*(pi/p))/l)
  return(gram)
}

```

## Calculate an Example value of the RBF Kernel 

Choose l = 2

```{r gram_matrix_user}
# With user defined RBF Kernel
l = 2
user_kernel_rbf_2 = rbf_kernel(X_scale_center,l)

# Convert to data frame for plotting
user_kernel_rbf_2_df = user_kernel_rbf_2 %>%
  as.data.frame() %>%
  mutate(xi = row_number()) %>%
  pivot_longer(starts_with("V"),names_to = "xj",values_to="Distance") %>%
  as.data.frame() %>%
  mutate(xj = as.numeric(str_remove(xj,"V")))

# Plot
user_kernel_rbf_2_plot = ggplot(data = user_kernel_rbf_2_df, aes(x=xi, y=xj, fill=Distance)) + 
  geom_raster() + 
  scale_fill_viridis_c() +
  scale_x_continuous(n.breaks = 10) +
  scale_y_continuous(n.breaks = 10) +
  labs(title = "User Defined RBF, l = 2, Kernel (Gram Matrix)") +
  theme_minimal()
```

## Calculate an Example value of the RBF Kernel Using Kernlab

Choosing l = 2. Note that the Kerlab package refers to this parameter as the "inverse kernel width." Trial and error appears to indicate this value is 1/l, as the RBF kernel is defined in this project (i.e. the parameter called "sigma" in the Kernlab documentation is in the numerator, not the demoninator)

```{r gram_matrix_kernlab}
## initialize kernel function 
l = 2
rbf = rbfdot(sigma = 1/l)

## calculate kernel matrix using Kernlab
kernlab_kernel_rbf_2 = kernelMatrix(rbf, X_scale_center)
# Convert to a data frame for plotting
kernlab_kernel_rbf_2_df = kernlab_kernel_rbf_2 %>%
  as.data.frame() %>%
  mutate(xi = row_number()) %>%
  pivot_longer(starts_with("V"),names_to = "xj",values_to="Distance") %>%
  as.data.frame() %>%
  mutate(xj = as.numeric(str_remove(xj,"V")))

# Plot
kernlab_gram_rbf_2_plot = ggplot(data = kernlab_kernel_rbf_2_df, aes(x=xi, y=xj, fill=Distance)) + 
  geom_raster() + 
  scale_fill_viridis_c() +
  scale_x_continuous(n.breaks = 10) +
  scale_y_continuous(n.breaks = 10) +
  labs(title = "Kernlab Defined RBF, sigma = .5, Kernel (Gram Matrix)") +
  theme_minimal()

```

## Confirm the User defined kernel matches the kernlab version

There are some rounding differences likely from computation, but the below confirms that the kernels from the kernlab package (rbf specification) and the user defined function are the same.

```{r gram_matrix_check}
# Check that the values are effectively 0
max(kernlab_kernel_rbf_2 - user_kernel_rbf_2)
min(kernlab_kernel_rbf_2 - user_kernel_rbf_2)

# Plot both side by side
# remove comments to export the visual
#png(filename = "Images/check_gram_rbf_2.png", width = 1080, height = 480)
plot_grid(user_kernel_rbf_2_plot,kernlab_gram_rbf_2_plot, labels = NULL, label_size = 12, ncol = 2, nrow = 1)
#dev.off()

remove(l, rbf, user_kernel_rbf_2_df, kernlab_kernel_rbf_2_df, user_kernel_rbf_2,
       kernlab_kernel_rbf_2, user_kernel_rbf_2_plot,kernlab_gram_rbf_2_plot)
```



## KPCA RBF Kernel: Exploring how Changes in l Effect Number of Principal Components

```{r kPCA_rbf}
## RBF, l = .01

## Note the kernlab package was used here. The "th" option  can be specificed in the kpca()
## function to allow for more PC and eigenvalues to be calculated. However, this section
## was sufficient in its currect form for the report and was left as is (e.g. th = 10^(-20))

# Note this section uses "sigma" and "l" interchangeably. The report sticks to "l"
# But some of the code was wirtten with "sigma" based on the definition in the 
# kernlab package


# sigma is defined as the "inverse kernel width on Kernlab", i.e. is in the denominator
kpca_kernlab_rbf_01 = kpca(~.,as.data.frame(X_scale_center),kernel="rbfdot",
            kpar=list(sigma=.01))

kpcs_rbf_pc_01 = pcv(kpca_kernlab_rbf_01)
kpcs_rbf_eig_01 = eig(kpca_kernlab_rbf_01)

# Percent of Variance Explained
percent_variance_explained_rbf_01 = 
  data.frame(PC = seq(1,length(kpcs_rbf_eig_01),by=1), 
             Percent_Variance = (kpcs_rbf_eig_01 / sum(kpcs_rbf_eig_01)))

# Create Scree Plot
kpca_rbf_scree_01 = ggplot(data = percent_variance_explained_rbf_01, aes(x = PC, y = Percent_Variance)) +
  geom_point() +
  geom_line() +
  labs(x = "Principal Component", y = "Percent of Variance Explained") +
  theme(text = element_text(size = 20),
    legend.title = element_text(size = 14),
    legend.text = element_text(size = 14),
    legend.key.width=unit(1,"cm"),
    axis.text.y = element_text(angle=90, hjust=1, size = 10),
    axis.text.x = element_text(size = 14),
    axis.title=element_text(size=14,face="bold"))


## RBF, Sigma = 1
# sigma is defined as the "inverse kernel width on Kernlab", i.e. is in the denominator
kpca_kernlab_rbf_1 = kpca(~.,as.data.frame(X_scale_center),kernel="rbfdot",
            kpar=list(sigma=1))

kpcs_rbf_pc_1 = pcv(kpca_kernlab_rbf_1)
kpcs_rbf_eig_1 = eig(kpca_kernlab_rbf_1)

# Percent of Variance Explained
percent_variance_explained_rbf_1 = 
  data.frame(PC = seq(1,length(kpcs_rbf_eig_1),by=1), 
             Percent_Variance = (kpcs_rbf_eig_1 / sum(kpcs_rbf_eig_1)))

# Create Scree Plot
kpca_rbf_scree_1 = ggplot(data = percent_variance_explained_rbf_1, aes(x = PC, y = Percent_Variance)) +
  geom_point() +
  geom_line() +
  labs(x = "Principal Component", y = "Percent of Variance Explained") +
  theme(text = element_text(size = 20),
    legend.title = element_text(size = 14),
    legend.text = element_text(size = 14),
    legend.key.width=unit(1,"cm"),
    axis.text.y = element_text(angle=90, hjust=1, size = 10),
    axis.text.x = element_text(size = 14),
    axis.title=element_text(size=14,face="bold"))

## RBF, Sigma = 2
# sigma is defined as the "inverse kernel width on Kernlab", i.e. is in the denominator
kpca_kernlab_rbf_2 = kpca(~.,as.data.frame(X_scale_center),kernel="rbfdot",
            kpar=list(sigma=2))

kpcs_rbf_pc_2 = pcv(kpca_kernlab_rbf_2)
kpcs_rbf_eig_2 = eig(kpca_kernlab_rbf_2)

# Percent of Variance Explained
percent_variance_explained_rbf_2 = 
  data.frame(PC = seq(1,length(kpcs_rbf_eig_2),by=1), 
             Percent_Variance = (kpcs_rbf_eig_2 / sum(kpcs_rbf_eig_2)))

# Create Scree Plot
kpca_rbf_scree_2 = ggplot(data = percent_variance_explained_rbf_2, aes(x = PC, y = Percent_Variance)) +
  geom_point() +
  geom_line() +
  labs(x = "Principal Component", y = "Percent of Variance Explained") +
  theme(text = element_text(size = 20),
    legend.title = element_text(size = 14),
    legend.text = element_text(size = 14),
    legend.key.width=unit(1,"cm"),
    axis.text.y = element_text(angle=90, hjust=1, size = 10),
    axis.text.x = element_text(size = 14),
    axis.title=element_text(size=14,face="bold"))


## Measure relationship between l and number of PCs
# required to get to 75% of variance explained

# Define a list of sigmas to explore (l in project formula)
sigma_list = seq(.01,2.1,by = .1)

# Specify dataframe to add the number of PCs (technically loading vectors)
# necessary
rbf_df = data.frame(sigma = sigma_list, pc_count = rep(NA,length(sigma_list)))

# Counter for loop
counter = 0

# Loop over each value
for (i in sigma_list) {
  counter = counter + 1
  
  # Perform KPCA with the specified value of sigma
  kpca_kernlab_rbf = kpca(~.,as.data.frame(X_scale_center),kernel="rbfdot",
              kpar=list(sigma=i))
  
  # Calculate the cumulative variance, as provided
  # by the normalized eigenvalues
  cumeig = cumsum(eig(kpca_kernlab_rbf)/sum(eig(kpca_kernlab_rbf)))
  
  # Get the number of eigenvalues (i.e. PCs) necessary
  # to get to 75%
  number = length(cumeig[which(cumeig < .75)]) + 1

  #store the variable
  rbf_df[counter,2] = as.numeric(number)
  
  remove(kpca_kernlab_rbf,cumeig,number)
}

# Plot the Value of sigma by the variable
pca_count_plot_rbf = ggplot(data = rbf_df, aes(x = sigma, y = pc_count)) + 
  geom_point() +
  geom_line() +
  labs(x = "l", y = "# of Principal Components \n needed to explain 75% variation") +
  theme(text = element_text(size = 20),
    legend.title = element_text(size = 14),
    legend.text = element_text(size = 14),
    legend.key.width=unit(1,"cm"),
    axis.text.y = element_text(angle=90, hjust=1, size = 10),
    axis.text.x = element_text(size = 14),
    axis.title=element_text(size=14,face="bold"))
  

# Save the output
# Remove comments to export the visual
#png(filename = "Images/rbf_varying_sigma.png", width = 960, height = 480)
pca_count_plot_rbf
#dev.off


# Plot scree plots side by side
# Remove comments to export the visual
#png(filename = "Images/rbf_sigma_01_and_2.png", width = 960, height = 480)
plot_grid(kpca_rbf_scree_01,kpca_rbf_scree_2, labels = NULL, label_size = 12, ncol = 2, nrow = 1)
#dev.off()

# Clean environment
remove(counter, i, kpcs_rbf_eig_01,kpcs_rbf_eig_1, kpcs_rbf_eig_2,
       sigma_list, kpca_kernlab_rbf_01, kpca_kernlab_rbf_1, 
       kpca_kernlab_rbf_2, kpca_rbf_scree_01, kpca_rbf_scree_1,
       kpca_rbf_scree_2,
       pca_count_plot_rbf, percent_variance_explained_rbf_01, 
       percent_variance_explained_rbf_1, percent_variance_explained_rbf_2, 
       rbf_df, kpcs_rbf_pc_01, kpcs_rbf_pc_1, kpcs_rbf_pc_2)
```

## KPCA RQ Kernel: Exploring Changes in l and alpha

```{r rq_kernel}
## Generate RBF Kernel and Plot
# RQ, l = .1, alpha = .01
rq_gram_1_01 = rq_kernel(X_scale_center,.1,.01)

# Convert to data frame for plotting
user_kernel_rq_1_01_df = rq_gram_1_01 %>%
  as.data.frame() %>%
  mutate(xi = row_number()) %>%
  pivot_longer(starts_with("V"),names_to = "xj",values_to="Distance") %>%
  as.data.frame() %>%
  mutate(xj = as.numeric(str_remove(xj,"V")))

# Plot
user_kernel_rq_1_01_plot = ggplot(data = user_kernel_rq_1_01_df, aes(x=xi, y=xj, fill=Distance)) + 
  geom_raster() + 
  scale_fill_viridis_c() +
  scale_x_continuous(n.breaks = 10) +
  scale_y_continuous(n.breaks = 10) +
  labs(title = "User Defined RQ, l = .1, alpha = .01 Kernel (Gram Matrix)") +
  theme_minimal()

# Export image for the report
# uncomment code to export the image
#png(filename = "Images/user_gram_rq_1_01.png", width = 480, height = 480)
user_kernel_rq_1_01_plot
#dev.off()



## Measure relationship between l and alpha and the number of PCs
# required to get to 75% of variance explained

# Define a list of l to explore
l_list = seq(.01,2.1,by = .1)
alpha_list = seq(.01,2.1,by = .1)

# Specify dataframe to add the number of PCs
# necessary
rq_df = data.frame(l = l_list, alpha = alpha_list)
rq_df = as.data.frame(expand(rq_df,l,alpha))
rq_df$pc_count = rep(NA,length(rq_df$l))

# Counter for loop
counter = 0

# Loop over each value
# Order matters here: depends on
# how df is constructed
for (l in l_list) {
  for (a in alpha_list) {
  counter = counter + 1
  
    # Calculate the Gram Matrix
    rq_gram = rq_kernel(X_scale_center,l,a)
    
    # Calculate the cumulative variance, as provided
    # by the normalized eigenvalues
    cumeig = cumsum(eigen(rq_gram)$values/sum(eigen(rq_gram)$values))
    
    # Get the number of eigenvalues (i.e. PCs) necessary
    # to get to 75%
    number = length(cumeig[which(cumeig < .75)]) + 1
  
    #store the variable
    rq_df[counter,3] = as.numeric(number)
    
    remove(rq_gram,cumeig,number)
  }
}

# Plot heatmap of l and alpha, with fill value being 
pca_count_plot_rq_log = ggplot(data = rq_df, aes(x=l, y=alpha, fill=log(pc_count))) + 
  geom_raster() + 
  scale_fill_viridis_c() +
  scale_x_continuous(n.breaks = 10) +
  scale_y_continuous(n.breaks = 10) +
  labs(title = "# of Principal Components \n needed to explain 75% variation") +
  theme_minimal()

# Plot heatmap of l and alpha, with fill value being 
pca_count_plot_rq = ggplot(data = rq_df, aes(x=l, y=alpha, fill=pc_count)) + 
  geom_raster() + 
  scale_fill_viridis_c() +
  scale_x_continuous(n.breaks = 10) +
  scale_y_continuous(n.breaks = 10) +
  labs(title = "# of Principal Components \n needed to explain 75% variation") +
  theme_minimal()
  

# Save the output
#png(filename = "Images/rq_varying_parameters.png", width = 1200, height = 480)
plot_grid(pca_count_plot_rq_log,pca_count_plot_rq, labels = NULL, label_size = 12, ncol = 2, nrow = 1)
#dev.off

# Clean environment
remove(pca_count_plot_rq, pca_count_plot_rq_log, rq_df, rq_gram_1_01, user_kernel_rq_1_01_df, 
       user_kernel_rq_1_01_plot,a, alpha_list, counter, l, l_list)
```


## KPCA PER Kernel: Exploring Changes in l and p

```{r rq_kernel}
## Generate RBF Kernel and Plot
# PER, l = .1, p = 24
per_gram_1_24 = per_kernel(X_scale_center,.1,24)

# Convert to data frame for plotting
user_kernel_per_1_24_df = per_gram_1_24 %>%
  as.data.frame() %>%
  mutate(xi = row_number()) %>%
  pivot_longer(starts_with("V"),names_to = "xj",values_to="Distance") %>%
  as.data.frame() %>%
  mutate(xj = as.numeric(str_remove(xj,"V")))

# Plot
user_kernel_per_1_24_plot = ggplot(data = user_kernel_per_1_24_df, aes(x=xi, y=xj, fill=Distance)) + 
  geom_raster() + 
  scale_fill_viridis_c() +
  scale_x_continuous(n.breaks = 10) +
  scale_y_continuous(n.breaks = 10) +
  labs(title = "User Defined PER, l = .1, p = 24 Kernel (Gram Matrix)") +
  theme_minimal()


# PER, l = .1, p = 7
per_gram_1_7 = per_kernel(X_scale_center,.1,7)

# Convert to data frame for plotting
user_kernel_per_1_7_df = per_gram_1_7 %>%
  as.data.frame() %>%
  mutate(xi = row_number()) %>%
  pivot_longer(starts_with("V"),names_to = "xj",values_to="Distance") %>%
  as.data.frame() %>%
  mutate(xj = as.numeric(str_remove(xj,"V")))

# Plot
user_kernel_per_1_7_plot = ggplot(data = user_kernel_per_1_7_df, aes(x=xi, y=xj, fill=Distance)) + 
  geom_raster() + 
  scale_fill_viridis_c() +
  scale_x_continuous(n.breaks = 10) +
  scale_y_continuous(n.breaks = 10) +
  labs(title = "User Defined PER, l = .1, p = 7 Kernel (Gram Matrix)") +
  theme_minimal()

# Export image for the report
# uncomment code to export the image
#png(filename = "Images/user_gram_per_01_7vs24.png", width = 1200, height = 480)
plot_grid(user_kernel_per_1_7_plot,user_kernel_per_1_24_plot, labels = NULL, label_size = 12, ncol = 2, nrow = 1)
#dev.off()




## Measure relationship between l and number of PCs
# required to get to 75% of variance explained
# fix p = 24

# Define a list of sigmas to explore (l in project formula)
l_list = seq(.01,2.1,by = .1)
p = 24

# Specify dataframe to add the number of PCs (technically loading vectors)
# necessary
per_df = data.frame(l = l_list, pc_count = rep(NA,length(l_list)))

# Counter for loop
counter = 0

# Loop over each value
for (i in l_list) {
  counter = counter + 1
  
    # Calculate the Gram Matrix
    per_gram = per_kernel(X_scale_center,i,p)
    
    # Calculate the cumulative variance, as provided
    # by the normalized eigenvalues
    cumeig = cumsum(eigen(per_gram)$values/sum(eigen(per_gram)$values))
    
    # Get the number of eigenvalues (i.e. PCs) necessary
    # to get to 75%
    number = length(cumeig[which(cumeig < .75)]) + 1
  
    #store the variable
    per_df[counter,2] = as.numeric(number)
    
    remove(per_gram,cumeig,number)
}

# Plot the Value of sigma by the variable
pca_count_plot_per = ggplot(data = per_df, aes(x = l, y = pc_count)) + 
  geom_point() +
  geom_line() +
  labs(x = "l", y = "# of Principal Components \n needed to explain 75% variation\np=24") +
  theme(text = element_text(size = 20),
    legend.title = element_text(size = 14),
    legend.text = element_text(size = 14),
    legend.key.width=unit(1,"cm"),
    axis.text.y = element_text(angle=90, hjust=1, size = 10),
    axis.text.x = element_text(size = 14),
    axis.title=element_text(size=14,face="bold"))
  

# Save the output
# Unccoment to output plots
#png(filename = "Images/per_varying_l_p_24.png", width = 960, height = 480)
pca_count_plot_per
#dev.off




## Scree Plot for l = .4, p = 24
# sigma is defined as the "inverse kernel width on Kernlab", i.e. is in the denominator
kpca_per_4_24 = per_kernel(X_scale_center,.4,24)

per_4_24_eigenvalues = eigen(kpca_per_4_24)$values

# Percent of Variance Explained
percent_variance_explained_per_4_24 = 
  data.frame(PC = seq(1,length(per_4_24_eigenvalues),by=1), 
             Percent_Variance = (per_4_24_eigenvalues / sum(per_4_24_eigenvalues)))

# Create Scree Plot
# Only show the first 30
kpca_per_scree_4_24 = ggplot(data = percent_variance_explained_per_4_24[1:30,], aes(x = PC, y = Percent_Variance)) +
  geom_point() +
  geom_line() +
  labs(x = "Principal Component", y = "Percent of Variance Explained") +
  theme(text = element_text(size = 20),
    legend.title = element_text(size = 14),
    legend.text = element_text(size = 14),
    legend.key.width=unit(1,"cm"),
    axis.text.y = element_text(angle=90, hjust=1, size = 10),
    axis.text.x = element_text(size = 14),
    axis.title=element_text(size=14,face="bold"))

# plot the output
# uncomment to output plots
#png(filename = "Images/per_4_24.png", width = 480, height = 480)
kpca_per_scree_4_24
#dev.off()

# clean environment
remove(kpca_per_4_24,kpca_per_scree_4_24, pca_count_plot_per, per_df, 
       per_gram_1_24, per_gram_1_7, percent_variance_explained_per_4_24, 
       user_kernel_per_1_24_df, user_kernel_per_1_24_plot, user_kernel_per_1_7_df,
       user_kernel_per_1_7_plot, counter, i, l_list, p, per_4_24_eigenvalues)
```

## Kernel Principal Component Regression

```{r kpcr}
# Perform SVD Using R Function
# To get the Principal components
kpca_per_4_24 = per_kernel(X_scale_center,.4,24)
K_per_svd = svd(kpca_per_4_24)

# Get Principal Compents by m
k_per_pca = K_per_svd$u%*%diag(K_per_svd$d)
  
# Fit model
kpcr_per = lm(Y_Matrix ~ k_per_pca[,c(1:6)] - 1)
summary(kpcr_per)

# Clean environment
remove(k_per_pca, K_per_svd, kpca_per_4_24)
```
